+++ .gitignore
@@ -4,8 +4,8 @@
 .idea/
 build/
 css/*.map
-js/polls.js
-js/polls.js.map
+js/*polls.*
+js/*.svg
 nbproject/
 node_modules/
 npm-debug.log
