+++ package.json
@@ -41,7 +41,8 @@
 		"stylelint-scss": "^3.4.4",
 		"v-tooltip": "^2.0.0-rc.33",
 		"vue": "^2.5.21",
-		"vue-click-outside": "^1.0.7"
+		"vue-click-outside": "^1.0.7",
+		"vue-router": "^3.0.2"
 	},
 	"browserslist": [
 		"> 1%",
@@ -77,7 +78,7 @@
 		"stylelint": "^9.9.0",
 		"stylelint-config-recommended-scss": "^3.2.0",
 		"stylelint-webpack-plugin": "^0.10.5",
-		"vue-loader": "^15.4.2",
+		"vue-loader": "^15.5.0",
 		"vue-style-loader": "^4.1.1",
 		"vue-template-compiler": "^2.5.21",
 		"webpack": "^4.28.2",
